further clean up of merge tree references
