+++ packages/dds/merge-tree/src/mergeTreeNodes.ts
@@ -16,10 +16,10 @@ import {
 import {
      LocalReferenceCollection,
 } from "./localReference";
+import { MergeTree } from "./mergeTree";
 import {
     IMergeTreeDeltaOpArgs,
 } from "./mergeTreeDeltaCallback";
-import { MergeTree } from "./mergeTree";
 import { TrackingGroupCollection } from "./mergeTreeTracking";
 import {
     ICombiningOp,
@@ -71,6 +71,10 @@ export interface IMergeBlock extends IMergeNodeCommon {
 
 export interface IHierBlock extends IMergeBlock {
     hierToString(indentCount: number): string;
+    /**
+     * @deprecated  for internal use only. public export will be removed.
+     * @internal
+     */
     addNodeReferences(mergeTree: MergeTree, node: IMergeNode): void;
     rightmostTiles: MapLike<ReferencePosition>;
     leftmostTiles: MapLike<ReferencePosition>;
@@ -128,6 +132,9 @@ export interface ISegment extends IMergeNodeCommon, Partial<IRemovalInfo> {
      * Returns true if the op  modifies the segment, otherwise false.
      * The only current false case is overlapping remove, where a segment is removed
      * by a previously sequenced operation before the current operation is acked.
+     *
+     * @deprecated  for internal use only. public export will be removed.
+     * @internal
      */
     ack(segmentGroup: SegmentGroup, opArgs: IMergeTreeDeltaOpArgs, mergeTree: MergeTree): boolean;
 }
@@ -363,7 +370,7 @@ export abstract class BaseSegment extends MergeNode implements ISegment {
 
     public abstract toJSONObject(): any;
 
-    public ack(segmentGroup: SegmentGroup, opArgs: IMergeTreeDeltaOpArgs, mergeTree: MergeTree): boolean {
+    public ack(segmentGroup: SegmentGroup, opArgs: IMergeTreeDeltaOpArgs): boolean {
         const currentSegmentGroup = this.segmentGroups.dequeue();
         assert(currentSegmentGroup === segmentGroup, 0x043 /* "On ack, unexpected segmentGroup!" */);
         switch (opArgs.op.type) {
