+++ packages/dds/merge-tree/src/test/beastTest.ts
@@ -43,11 +43,11 @@ import {
 import { SnapshotLegacy } from "../snapshotlegacy";
 import {
     IJSONTextSegment,
-    MergeTreeTextHelper,
     TextSegment,
 } from "../textSegment";
 import { reservedRangeLabelsKey, reservedTileLabelsKey } from "../referencePositions";
 import { MergeTree } from "../mergeTree";
+import { MergeTreeTextHelper } from "../MergeTreeTextHelper";
 import { specToSegment, TestClient } from "./testClient";
 import { TestServer } from "./testServer";
 import { insertText, loadTextFromFile, nodeOrdinalsHaveIntegrity } from "./testUtils";
