+++ packages/dds/merge-tree/src/test/testServer.ts
@@ -18,7 +18,8 @@ import {
 } from "../mergeTreeNodes";
 import { ClientSeq, clientSeqComparer } from "../mergeTree";
 import { PropertySet } from "../properties";
-import { MergeTreeTextHelper, TextSegment } from "../textSegment";
+import { TextSegment } from "../textSegment";
+import { MergeTreeTextHelper } from "../MergeTreeTextHelper";
 import { TestClient } from "./testClient";
 
 /**
